December 30th, 2025 Update

+++ opendbc_repo/opendbc/safety/modes/hyundai_common.h
@@ -49,13 +49,13 @@ bool hyundai_has_lda_button = false;
 static uint8_t hyundai_last_button_interaction;  // button messages since the user pressed an enable button
 
 void hyundai_common_init(uint16_t param) {
-  const int HYUNDAI_PARAM_EV_GAS = 1;
-  const int HYUNDAI_PARAM_HYBRID_GAS = 2;
-  const int HYUNDAI_PARAM_CAMERA_SCC = 8;
-  const int HYUNDAI_PARAM_CANFD_LKA_STEERING = 16;
-  const int HYUNDAI_PARAM_ALT_LIMITS = 64; // TODO: shift this down with the rest of the common flags
-  const int HYUNDAI_PARAM_FCEV_GAS = 256;
-  const int HYUNDAI_PARAM_ALT_LIMITS_2 = 512;
+  const uint16_t HYUNDAI_PARAM_EV_GAS = 1;
+  const uint16_t HYUNDAI_PARAM_HYBRID_GAS = 2;
+  const uint16_t HYUNDAI_PARAM_CAMERA_SCC = 8;
+  const uint16_t HYUNDAI_PARAM_CANFD_LKA_STEERING = 16;
+  const uint16_t HYUNDAI_PARAM_ALT_LIMITS = 64; // TODO: shift this down with the rest of the common flags
+  const uint16_t HYUNDAI_PARAM_FCEV_GAS = 256;
+  const uint16_t HYUNDAI_PARAM_ALT_LIMITS_2 = 512;
 
   // FrogPilot variables
   const int HYUNDAI_PARAM_HAS_LDA_BUTTON = 1024;
@@ -74,7 +74,7 @@ void hyundai_common_init(uint16_t param) {
   hyundai_last_button_interaction = HYUNDAI_PREV_BUTTON_SAMPLES;
 
 #ifdef ALLOW_DEBUG
-  const int HYUNDAI_PARAM_LONGITUDINAL = 4;
+  const uint16_t HYUNDAI_PARAM_LONGITUDINAL = 4;
   hyundai_longitudinal = GET_FLAG(param, HYUNDAI_PARAM_LONGITUDINAL);
 #else
   hyundai_longitudinal = false;
